renamed mod optimizer to mod optimizers and do not re-export losses
