+++ mlx-nn/src/lib.rs
@@ -5,8 +5,9 @@
 //! All modules provide a `new()` function that take mandatory parameters and other methods
 //! to set optional parameters.
 
+pub mod losses;
 pub mod macros;
-pub mod optimizer;
+pub mod optimizers;
 pub mod utils;
 
 mod activation;
@@ -14,7 +15,6 @@ mod container;
 mod convolution;
 mod dropout;
 mod linear;
-mod losses;
 mod value_and_grad;
 
 pub use activation::*;
@@ -22,7 +22,6 @@ pub use container::*;
 pub use convolution::*;
 pub use dropout::*;
 pub use linear::*;
-pub use losses::*;
 pub use value_and_grad::*;
 
 /// Re-export of the `mlx-nn-module` crate.
